aheui: store the list element count as u32 and declare its width to the JIT
`Stack`/`Queue`/`Port` hold `size: u32` instead of `usize`; the `LinkedList`
accessors convert at the boundary so the trait keeps its `usize` contract, and
the struct sizes and field offsets are unchanged. The `1usize` literals in the
depth arithmetic become `1u32`.

The `jit_inline` and `jit_interp` sites declare
`int_fields = { Stack::size => u32, Queue::size => u32, Port::size => u32 }`.
All three lists share the `head`/`size` prefix and the JIT reaches them through
the `Stack` tag, so each name is declared. The emitted `.size` loads are 4-byte
unsigned, and `intbounds` bounds them: on logo the compiled trace goes from
8925 emits + 2514 guards to 3830 + 610, of which `GuardTrue` is 2209 -> 305.

+++ aheui-runtime/src/storage/mod.rs
@@ -1110,7 +1110,7 @@ impl Storage {
                     count += 1;
                     cur = unsafe { (*cur).next };
                 }
-                stack.size = count;
+                stack.size = count as u32;
             }
         }
     }
